+++ user_guide_src/source/changelog.rst
@@ -83,6 +83,7 @@ Release Date: Not Released
 	 - Added an optional paramater to ``delete_files()`` to enable it to skip deleting files such as .htaccess and index.html.
 	 - ``read_file()`` is now a deprecated alias of ``file_get_contents()``.
    -  :doc:`Security Helper <helpers/security_helper>` function ``strip_image_tags()`` is now an alias for the same method in the :doc:`Security Library <libraries/security>`.
+   -  :doc:`Form Helper <helpers/form_helper>` function ``form_hidden()`` added an optional parameter to add attributes.
 
 -  Database